Reorder workshop (separate out data tables from scanning recipe)
Separate out data tables from scanning recipe so we can cover data tables first
